uplink: move `ContractError` to its own module
