+++ piecrust-uplink/src/lib.rs
@@ -12,7 +12,7 @@ extern crate alloc;
 mod state;
 pub use state::{
     call, call_raw, call_raw_with_limit, call_with_limit, caller, emit,
-    host_query, limit, meta_data, owner, self_id, spent, ContractError,
+    host_query, limit, meta_data, owner, self_id, spent,
 };
 
 mod helpers;
@@ -20,6 +20,9 @@ pub use helpers::*;
 
 mod types;
 pub use types::*;
+
+mod error;
+pub use error::*;
 
 pub mod bufwriter;
 pub mod debug;
